# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: KSA Attendance System API
  version: 0.2.0
  summary: September 2026 implementation baseline between the production frontend and backend.
  description: |
    This contract captures approved API capabilities and stable domain outcomes for
    implementation. Endpoint names may change through a documented superseding decision, but the
    authorization, state, privacy, idempotency, and source-of-truth guarantees must remain.

    The permanent QR points to a stable frontend route such as `/check-in`; it is not an
    API credential. The frontend obtains an explicit browser location and submits it to
    the backend, which makes the final attendance decision using server time.

    The production browser uses the same public origin for the UI and `/api/*`. Cookie-authenticated
    state changes also require the implementation's CSRF/origin control. Abuse-sensitive operations
    may return the documented `RATE_LIMITED` response and safe `Retry-After` under DEC-036.
servers:
  - url: /api/v1
    description: Same-origin API
...
security:
  - sessionCookie: []
paths:
...
  /me/device:
    get:
      tags: [Devices]
      summary: Return safe current-browser and registered attendance-device status
      responses:
        "200":
          description: Device status; raw credentials are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceStatusEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
  /me/device-change-requests:
    get:
      tags: [Devices]
      summary: List the current participant's own device-change requests
      responses:
        "200":
          description: Own requests
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeviceChangeRequest"
        "403":
          $ref: "#/components/responses/AuthorizationError"
    post:
      tags: [Devices]
      summary: Request approval for the current browser as a replacement attendance device
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 500
      responses:
        "201":
          description: Request created or current pending request returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
...
  /attendance/context:
    get:
      tags: [Attendance]
      summary: Return the current participant-safe attendance state for the stable check-in route
      responses:
        "200":
          description: Current attendance context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceContextEnvelope"
        "401":
          $ref: "#/components/responses/AuthenticationError"
  /attendance/check-in:
    post:
      tags: [Attendance]
      summary: Evaluate and, when valid, atomically record current participant attendance
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: Deterministic attendance outcome, including duplicate success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResultEnvelope"
        "201":
          description: New attendance record committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResultEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          description: Session or attendance-device state prevents automatic check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResultEnvelope"
        "422":
          description: Location could not be automatically accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResultEnvelope"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/UnavailableError"
...
  /device-change-requests:
    get:
      tags: [Devices]
      summary: List device-change requests visible to an operational reviewer
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReviewStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeviceChangeRequest"
                  nextCursor:
                    type: [string, "null"]
        "403":
          $ref: "#/components/responses/AuthorizationError"
  /device-change-requests/{requestId}:
    get:
      tags: [Devices]
      summary: Get an authorized device-change review including protected identity-photo access
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Device-change request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /device-change-requests/{requestId}/approve:
    post:
      tags: [Devices]
      summary: Approve candidate device and atomically revoke the previous active device
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Approved or already-final request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /device-change-requests/{requestId}/reject:
    post:
      tags: [Devices]
      summary: Reject a candidate attendance device
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Rejected or already-final request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceChangeRequestEnvelope"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ksa_session
      description: Secure, HttpOnly, host-only, SameSite=Lax server session; raw value is never stored. Idle expiry is 30 days and absolute expiry is 90 days.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthenticationError:
      description: Authentication is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthorizationError:
      description: Current actor is not allowed to perform this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: Resource is absent or intentionally undisclosed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: Current domain state conflicts with the requested command
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimitedError:
      description: Too many requests; retry according to safe response guidance
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnavailableError:
      description: Authoritative service unavailable before a valid commit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Role:
      type: string
      enum: [PARTICIPANT, COURSE_REP, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
...
    ParticipantSummary:
      type: object
      required: [id, fullName, serialNumber, status, roles]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        serialNumber:
          type: string
          pattern: "^KSA-[0-9]{2}$"
        status:
          $ref: "#/components/schemas/UserStatus"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    ActorSummary:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
...
    SessionStatus:
      type: string
      enum: [DRAFT, SCHEDULED, OPEN, CLOSED, CANCELLED]
    AttendanceSession:
      type: object
      required:
        [
          id,
          attendanceDate,
          effectiveStart,
          originalEffectiveEnd,
          effectiveEnd,
          timezone,
          status,
          version,
          createdAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        attendanceDate:
          type: string
          format: date
        effectiveStart:
          type: string
          format: date-time
        effectiveEnd:
          type: string
          format: date-time
        originalEffectiveEnd:
          type: string
          format: date-time
          description: Immutable baseline used for the Course Rep two-hour extension cap
        timezone:
          type: string
          example: Africa/Lagos
        status:
          $ref: "#/components/schemas/SessionStatus"
        version:
          type: integer
          minimum: 1
        openedAt:
          type: [string, "null"]
          format: date-time
        openedBy:
          type: [string, "null"]
        closedAt:
          type: [string, "null"]
          format: date-time
        closedBy:
          type: [string, "null"]
          description: User ID or SYSTEM marker
        cancelledAt:
          type: [string, "null"]
          format: date-time
        extendedAt:
          type: [string, "null"]
          format: date-time
        extendedBy:
          type: [string, "null"]
        reopenedAt:
          type: [string, "null"]
          format: date-time
        reopenCount:
          type: integer
          minimum: 0
        presentCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
...
    LocationReading:
      type: object
      required: [latitude, longitude, accuracyMetres, capturedAt]
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracyMetres:
          type: number
          exclusiveMinimum: 0
        capturedAt:
          type: string
          format: date-time
          description: Fixes older than 30 seconds are discarded. Coordinates are used in-request only and are never persisted.
    CheckInRequest:
      type: object
      properties:
        location:
          $ref: "#/components/schemas/LocationReading"
        locationFailure:
          type: string
          enum: [PERMISSION_DENIED, TIMEOUT, UNAVAILABLE, UNSUPPORTED]
      oneOf:
        - required: [location]
        - required: [locationFailure]
      additionalProperties: false
    CheckInOutcome:
      type: string
      enum:
        - ATTENDANCE_RECORDED
        - ALREADY_CHECKED_IN
        - SESSION_NOT_OPEN
        - SESSION_CLOSED
        - DEVICE_CHANGE_REQUIRED
        - LOCATION_UNCERTAIN
        - LOCATION_PERMISSION_DENIED
        - LOCATION_TIMEOUT
        - LOCATION_UNAVAILABLE
        - LOCATION_UNSUPPORTED
        - CLEARLY_REMOTE
    CheckInResult:
      type: object
      required: [outcome, serverTime, manualRequestAllowed]
      properties:
        outcome:
          $ref: "#/components/schemas/CheckInOutcome"
        serverTime:
          type: string
          format: date-time
        attemptId:
          type: [string, "null"]
          format: uuid
          description: Safe reference used for an explicit manual-verification request when allowed
        manualRequestAllowed:
          type: boolean
          description: True only for a location outcome eligible for explicit manual review
        attendanceRecord:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
        manualVerificationCase:
          oneOf:
            - $ref: "#/components/schemas/ManualVerificationCase"
            - type: "null"
        deviceChangeRequest:
          oneOf:
            - $ref: "#/components/schemas/DeviceChangeRequest"
            - type: "null"
    AttendanceMethod:
      type: string
      enum: [QR, MANUAL]
...
    AttendanceRecord:
      type: object
      required: [id, sessionId, participant, status, method, checkedInAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        status:
          type: string
          enum: [PRESENT]
        method:
          $ref: "#/components/schemas/AttendanceMethod"
        checkedInAt:
          type: string
          format: date-time
        approvedBy:
          oneOf:
            - $ref: "#/components/schemas/ActorSummary"
            - type: "null"
        locationStatus:
          type: string
          enum: [VERIFIED, MANUAL, NOT_RETAINED]
...
    AttendanceContext:
      type: object
      required: [state, deviceStatus]
      properties:
        state:
          type: string
          enum: [NOT_OPEN, OPEN, CLOSED, ALREADY_PRESENT]
        session:
          oneOf:
            - $ref: "#/components/schemas/AttendanceSession"
            - type: "null"
        attendanceRecord:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
        deviceStatus:
          type: string
          enum: [REGISTERED_BROWSER, UNRECOGNIZED_BROWSER, NO_ACTIVE_DEVICE]
    ReviewStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED, SUPERSEDED]
    ManualVerificationCase:
      type: object
      required:
        [
          id,
          participant,
          sessionId,
          creationSource,
          reasonCode,
          status,
          requestedAt,
          expiresAt,
          version,
        ]
      properties:
        id:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        sessionId:
          type: string
          format: uuid
        sourceAttemptId:
          type: [string, "null"]
          format: uuid
        creationSource:
          type: string
          enum: [AUTOMATIC_UNCERTAIN, PARTICIPANT_REQUEST, OPERATOR_EMERGENCY]
        reasonCode:
          type: string
          enum:
            [
              LOCATION_UNCERTAIN,
              LOCATION_PERMISSION_DENIED,
              LOCATION_TIMEOUT,
              LOCATION_UNAVAILABLE,
              LOCATION_UNSUPPORTED,
              CLEARLY_REMOTE,
              EMERGENCY,
            ]
        participantReason:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ReviewStatus"
        requestedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          description: Session close plus 15 minutes
        reviewedAt:
          type: [string, "null"]
          format: date-time
        reviewedBy:
          type: [string, "null"]
          format: uuid
        decisionReason:
          type: [string, "null"]
        photoAccessUrl:
          type: [string, "null"]
          description: Authorized presigned URL lasting only a few minutes; detail response only
        attendanceRecord:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
        version:
          type: integer
          minimum: 1
    ReviewDecisionRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 3
          maxLength: 1000
    DeviceChangeRequest:
      type: object
      required: [id, participant, status, requestedAt, version]
      properties:
        id:
          type: string
          format: uuid
        participant:
          $ref: "#/components/schemas/ParticipantSummary"
        status:
          $ref: "#/components/schemas/ReviewStatus"
        requestedAt:
          type: string
          format: date-time
        reviewedAt:
          type: [string, "null"]
          format: date-time
        reviewedBy:
          type: [string, "null"]
          format: uuid
        decisionReason:
          type: [string, "null"]
        photoAccessUrl:
          type: [string, "null"]
          description: Authorized short-lived or proxied URL; detail response only
        version:
          type: integer
          minimum: 1
...
    DeviceStatus:
      type: object
      required: [browserStatus]
      properties:
        browserStatus:
          type: string
          enum: [REGISTERED_BROWSER, UNRECOGNIZED_BROWSER, NO_ACTIVE_DEVICE]
        activeDeviceApprovedAt:
          type: [string, "null"]
          format: date-time
        pendingRequest:
          oneOf:
            - $ref: "#/components/schemas/DeviceChangeRequest"
            - type: "null"
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - STALE_STATE
            - DUPLICATE_EMAIL
            - DUPLICATE_PHONE
            - DUPLICATE_SERIAL
            - ROSTER_ENTRY_NOT_ELIGIBLE
            - ROSTER_CLAIM_CONFLICT
            - LAST_ADMIN_REQUIRED
            - RECENT_REAUTHENTICATION_REQUIRED
            - SELF_APPROVAL_FORBIDDEN
            - SESSION_ALREADY_OPEN
            - SESSION_NOT_OPEN
            - SESSION_CLOSED
            - SESSION_EXTENSION_LIMIT
            - DEVICE_CHANGE_REQUIRED
            - LOCATION_UNCERTAIN
            - LOCATION_PERMISSION_DENIED
            - LOCATION_TIMEOUT
            - LOCATION_UNAVAILABLE
            - LOCATION_UNSUPPORTED
            - CLEARLY_REMOTE
            - MANUAL_VERIFICATION_NOT_ALLOWED
            - MANUAL_VERIFICATION_EXPIRED
            - PHOTO_TOO_LARGE
            - PHOTO_UNSUPPORTED
            - RATE_LIMITED
            - SERVICE_TEMPORARILY_UNAVAILABLE
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    AttendanceContextEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AttendanceContext"
    CheckInResultEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CheckInResult"
...
    DeviceChangeRequestEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/DeviceChangeRequest"
...
    DeviceStatusEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/DeviceStatus"
...